[#45] SearchResultView에 정의된 struct를 파일로 분리시키기
#45
